+++ packages/blocks/core/CHANGELOG.md
@@ -80,6 +80,7 @@
 - `Archive` block: support improved.
 - `Audio` block: support improved.
 - `Avatar` block: support improved.
+- `Button` block: support improved.
 
 
 ### Automated Tests
@@ -97,6 +98,7 @@
 - Updated E2E tests to check support of `Archive` block.
 - Updated E2E tests to check support of `Audio` block.
 - Updated E2E tests to check support of `Avatar` block.
+- Updated E2E tests to check support of `Button` block.
 
 
 ## 1.3.0 (2025-04-27)